fix: don't split completions that aren't file paths
We were previously indiscriminately removing all but the last
component of path-like completion results when displaying them
for selection (but inserting the full string). We should only
do that when we know the completions are supposed to be processed
as file paths.

This fixes, among other things, completion display of git refs
when completing on `git checkout `.