fix: Posix compliant argument parsing for -c
- clap does not handle `--` as an argument but rather as a separator so we need additional processing
for passing `--` as an operand to a `-c` script. This is because Posix interprets `--` as an operand. The first argument in `-c` is the name of the command and can be `--` just fine.

See: Posix about sh
     https://pubs.opengroup.org/onlinepubs/9699919799/utilities/sh.html
     https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap12.html#tag_12_02

test: tests for command line argument handling

feat: helper functions to deal with clap's limitation handling `--`

- Added functions `try_parse_known` and `parse_known` for splitting arguments into `before hyphen` and `hyphen and after hyphen`
- This approach allows manually deciding what to do with `--` and the remaining arguments

Fixes:
- main CommandLineArgs
- EchoCommand

Clap issue: https://github.com/clap-rs/clap/issues/5055

fix: clippy lints